requirements.txt
requests
setproctitle
redis
bcrypt
flask
flask-sqlalchemy
flask-session
flask-login
gunicorn
python-telegram-bot
python-dotenv